Made delete button functional
