--- src/components/comments-table/comments-line.module.css
@@ -1,19 +1,0 @@
-.line {
-  padding-top: 12px;
-  padding-bottom: 12px;
-  border-bottom: 1px solid #83828F;
-  position: relative;
-}
-
-.button_container {
-  text-align: end;
-}
-
-.button {
-  width: 21px;
-  height: 21px;
-  border: none;
-  cursor: pointer;
-  padding: 0;
-  background-size: cover;
-}
